+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/RawCustomResourceOperationsImpl.java
@@ -35,6 +35,14 @@ import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 
+/**
+ * This class simple does basic operations for custom defined resources without
+ * demanding the POJOs for custom resources. It is serializing/deserializing
+ * objects to plain hash map(String, Object).
+ *
+ * Right now it supports basic operations like GET, POST, PUT, DELETE.
+ *
+ */
 public class RawCustomResourceOperationsImpl extends OperationSupport {
   private OkHttpClient client;
   private Config config;
@@ -50,110 +58,318 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
     this.objectMapper = Serialization.jsonMapper();
   }
 
+  /**
+   * Load a custom resource object from an inputstream into a HashMap
+   *
+   * @param fileInputStream file input stream
+   * @return custom resource as HashMap
+   * @throws IOException exception in case any read operation fails.
+   */
   public Map<String, Object> load(InputStream fileInputStream) throws IOException {
     return convertJsonStringToMap(IOHelpers.readFully(fileInputStream));
   }
 
+  /**
+   * Load a custom resource object from a JSON string into a HashMap
+   *
+   * @param objectAsJsonString object as JSON string
+   * @return custom resource as HashMap
+   * @throws IOException exception in case any problem in reading json.
+   */
   public Map<String, Object> load(String objectAsJsonString) throws IOException {
     return convertJsonStringToMap(objectAsJsonString);
   }
 
+  /**
+   * Create a custom resource which is a non-namespaced object.
+   *
+   * @param objectAsString object as JSON string
+   * @return Object as HashMap
+   * @throws IOException exception in case of any network/read problems
+   */
   public Map<String, Object> create(String objectAsString) throws IOException {
     return validateAndSubmitRequest(null, null, objectAsString, HttpCallMethod.POST);
   }
 
+  /**
+   * Create a custom resource which is non-namespaced.
+   *
+   * @param object object a HashMap
+   * @return Object as HashMap
+   * @throws KubernetesClientException in case of error from Kubernetes API
+   * @throws IOException in case of problems while reading HashMap
+   */
   public Map<String, Object> create(Map<String, Object> object) throws KubernetesClientException, IOException {
     return validateAndSubmitRequest(null, null, objectMapper.writeValueAsString(object), HttpCallMethod.POST);
   }
 
+  /**
+   * Create a custom resource which is a namespaced object.
+   *
+   * @param namespace namespace in which we want object to be created.
+   * @param objectAsString Object as JSON string
+   * @return Object as HashMap
+   * @throws KubernetesClientException in case of error from Kubernetes API
+   * @throws IOException in case of problems while reading JSON object
+   */
   public Map<String, Object> create(String namespace, String objectAsString) throws KubernetesClientException, IOException {
     return validateAndSubmitRequest(namespace, null, objectAsString, HttpCallMethod.POST);
   }
 
+  /**
+   * Create a custom resource which is non-namespaced object.
+   *
+   * @param objectAsStream object as a file input stream
+   * @return Object as HashMap
+   * @throws KubernetesClientException in case of error from Kubernetes API
+   * @throws IOException in case of problems while reading file
+   */
   public Map<String, Object> create(InputStream objectAsStream) throws KubernetesClientException, IOException {
     return validateAndSubmitRequest(null, null, IOHelpers.readFully(objectAsStream), HttpCallMethod.POST);
   }
 
+  /**
+   * Create a custom resource which is a namespaced object.
+   *
+   * @param namespace namespace in which we want object to be created
+   * @param objectAsStream object as file input stream
+   * @return Object as HashMap
+   * @throws KubernetesClientException in case of error from Kubernetes API
+   * @throws IOException in case of problems while reading file
+   */
   public Map<String, Object> create(String namespace, InputStream objectAsStream) throws KubernetesClientException, IOException {
     return validateAndSubmitRequest(namespace, null, IOHelpers.readFully(objectAsStream), HttpCallMethod.POST);
   }
 
+  /**
+   * Create a custom resource which is a namespaced object.
+   *
+   * @param namespace namespace in which we want object to be created.
+   * @param object object as a HashMap
+   * @return Object as HashMap
+   * @throws KubernetesClientException in case of error from Kubernetes API
+   * @throws IOException in case of problems faced while serializing HashMap
+   */
   public Map<String, Object> create(String namespace, Map<String, Object> object) throws KubernetesClientException, IOException {
     return validateAndSubmitRequest(namespace, null, objectMapper.writeValueAsString(object), HttpCallMethod.POST);
   }
 
+  /**
+   *
+   * Create or replace a custom resource which is a non-namespaced object.
+   *
+   * @param objectAsString object as JSON string
+   * @return Object as HashMap
+   * @throws IOException in case of network/serializiation failures or failures from Kuberntes API
+   */
   public Map<String, Object> createOrReplace(String objectAsString) throws IOException {
     return createOrReplaceJsonStringObject(null, objectAsString);
   }
 
+  /**
+   * Create or replace a custom resource which is a non-namespced object.
+   *
+   * @param customResourceObject object as HashMap
+   * @return Object as HashMap
+   * @throws IOException in case of network/serialization failures or failures from Kubernetes API
+   */
   public Map<String, Object> createOrReplace(Map<String, Object> customResourceObject) throws IOException {
     return createOrReplace(objectMapper.writeValueAsString(customResourceObject));
   }
 
+  /**
+   * Create or replace a custom resource which is non-namespaced object.
+   *
+   * @param inputStream object as file input stream
+   * @return Object as HashMap
+   * @throws IOException in case of network/serialization failures or failures from Kubernetes API
+   */
   public Map<String, Object> createOrReplace(InputStream inputStream) throws IOException {
     return createOrReplace(IOHelpers.readFully(inputStream));
   }
 
+  /**
+   * Create or replace a custom resource which is namespaced object.
+   *
+   * @param namespace desired namespace
+   * @param objectAsString object as JSON String
+   * @return Object as HashMap
+   * @throws IOException in case of network/serialization failures or failures from Kubernetes API
+   */
   public Map<String, Object> createOrReplace(String namespace, String objectAsString) throws IOException {
     return createOrReplaceJsonStringObject(namespace, objectAsString);
   }
 
+  /**
+   * Create or replace a custom resource which is namespaced object.
+   *
+   * @param namespace desired namespace
+   * @param customResourceObject object as HashMap
+   * @return Object as HashMap
+   * @throws IOException in case of network/serialization failures or failures from Kubernetes API
+   */
   public Map<String, Object> createOrReplace(String namespace, Map<String, Object> customResourceObject) throws IOException {
     return createOrReplace(namespace, objectMapper.writeValueAsString(customResourceObject));
   }
 
+  /**
+   * Create or replace a custom resource which is namespaced object.
+   *
+   * @param namespace desired namespace
+   * @param objectAsString object as file input stream
+   * @return Object as HashMap
+   * @throws IOException in case of network/serialization failures or failures from Kubernetes API
+   */
   public Map<String, Object> createOrReplace(String namespace, InputStream objectAsString) throws IOException {
     return createOrReplace(namespace, IOHelpers.readFully(objectAsString));
   }
 
+  /**
+   * Edit a custom resource object which is a non-namespaced object.
+   *
+   * @param name name of the custom resource
+   * @param object new object as a HashMap
+   * @return Object as HashMap
+   * @throws IOException in case of network/serialization failures or failures from Kubernetes API
+   */
   public Map<String, Object> edit(String name, Map<String, Object> object) throws IOException {
     return validateAndSubmitRequest(null, name, objectMapper.writeValueAsString(object), HttpCallMethod.PUT);
   }
 
+  /**
+   * Edit a custom resource object which is a non-namespaced object.
+   *
+   * @param name name of the custom resource
+   * @param objectAsString new object as a JSON String
+   * @return Object as HashMap
+   * @throws IOException in case of network/serialization failures or failures from Kubernetes API
+   */
   public Map<String, Object> edit(String name, String objectAsString) throws IOException {
     return validateAndSubmitRequest(null, name, objectAsString, HttpCallMethod.PUT);
   }
 
+  /**
+   * Edit a custom resource object which is a namespaced object.
+   *
+   * @param namespace desired namespace
+   * @param name name of the custom resource
+   * @param object new object as a HashMap
+   * @return Object as HashMap
+   * @throws IOException in case of network/serialization failures or failures from Kubernetes API
+   */
   public Map<String, Object> edit(String namespace, String name, Map<String, Object> object) throws IOException {
     return validateAndSubmitRequest(namespace, name, objectMapper.writeValueAsString(object), HttpCallMethod.PUT);
   }
 
+  /**
+   * Edit a custom resource object which is a namespaced object.
+   *
+   * @param namespace desired namespace
+   * @param name name of the custom resource
+   * @param objectAsString new object as a JSON string
+   * @return Object as HashMap
+   * @throws IOException in case of network/serialization failures or failures from Kubernetes API
+   */
   public Map<String, Object> edit(String namespace, String name, String objectAsString) throws IOException {
+    // Append resourceVersion in object metadata in order to
+    // avoid : https://github.com/fabric8io/kubernetes-client/issues/1724
+    objectAsString = appendResourceVersionInObject(namespace, name, objectAsString);
     return validateAndSubmitRequest(namespace, name, objectAsString, HttpCallMethod.PUT);
   }
 
+  /**
+   * Edit a custom resource object which is a non-namespaced object.
+   *
+   * @param name name of the custom resource
+   * @param objectAsStream new object as a file input stream
+   * @return Object as HashMap
+   * @throws IOException in case of network/serialization failures or failures from Kubernetes API
+   */
   public Map<String, Object> edit(String name, InputStream objectAsStream) throws IOException, KubernetesClientException {
     return validateAndSubmitRequest(null, name, IOHelpers.readFully(objectAsStream), HttpCallMethod.PUT);
   }
 
+  /**
+   * Edit a custom resource object which is a namespaced object.
+   *
+   * @param namespace desired namespace
+   * @param name name of the custom resource
+   * @param objectAsStream new object as a file input stream
+   * @return Object as HashMap
+   * @throws IOException in case of network/serialization failures or failures from Kubernetes API
+   */
   public Map<String, Object> edit(String namespace, String name, InputStream objectAsStream) throws IOException, KubernetesClientException {
     return validateAndSubmitRequest(namespace, name, IOHelpers.readFully(objectAsStream), HttpCallMethod.PUT);
   }
 
+  /**
+   * Get a custom resource from the cluster which is non-namespaced.
+   *
+   * @param name name of custom resource
+   * @return Object as HashMap
+   */
   public Map<String, Object> get(String name) {
     return makeCall(fetchUrl(null, null) + name, null, HttpCallMethod.GET);
   }
 
+  /**
+   * Get a custom resource from the cluster which is namespaced.
+   *
+   * @param namespace desired namespace
+   * @param name name of custom resource
+   * @return Object as HashMap
+   */
   public Map<String, Object> get(String namespace, String name) {
       return makeCall(fetchUrl(namespace, null) + name, null, HttpCallMethod.GET);
   }
 
+  /**
+   * List all custom resources in all namespaces
+   *
+   * @return list of custom resources as HashMap
+   */
   public Map<String, Object> list() {
     return makeCall(fetchUrl(null, null), null, HttpCallMethod.GET);
   }
 
+  /**
+   * List all custom resources in a specific namespace
+   *
+   * @param namespace desired namespace
+   * @return list of custom resources as HashMap
+   */
   public Map<String, Object> list(String namespace) {
     return makeCall(fetchUrl(namespace, null), null, HttpCallMethod.GET);
   }
 
+  /**
+   * List all custom resources in a specific namespace with some labels
+   *
+   * @param namespace desired namespace
+   * @param labels labels as a HashMap
+   * @return list of custom resources as HashMap
+   */
   public Map<String, Object> list(String namespace, Map<String, String> labels) {
     return makeCall(fetchUrl(namespace, labels), null, HttpCallMethod.GET);
   }
 
+  /**
+   * Delete all custom resources in a specific namespace
+   *
+   * @param namespace desired namespace
+   * @return deleted objects as HashMap
+   */
   public Map<String, Object> delete(String namespace) {
     return makeCall(fetchUrl(namespace, null), null, HttpCallMethod.DELETE);
   }
 
+  /**
+   * Delete a custom resource in a specific namespace
+   *
+   * @param namespace desired namespace
+   * @param name custom resource's name
+   * @return object as HashMap
+   */
   public Map<String, Object> delete(String namespace, String name) {
     return makeCall(fetchUrl(namespace, null) + name, null, HttpCallMethod.DELETE);
   }
@@ -266,5 +482,15 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
         return requestBuilder.put(requestBody).url(url).build();
     }
     return requestBuilder.build();
+  }
+
+  private String appendResourceVersionInObject(String namespace, String customResourceName, String customResourceAsJsonString) throws IOException {
+    Map<String, Object> oldObject = get(namespace, customResourceName);
+    String resourceVersion = ((Map<String, Object>)oldObject.get("metadata")).get("resourceVersion").toString();
+
+    Map<String, Object> newObject = convertJsonStringToMap(customResourceAsJsonString);
+    ((Map<String, Object>)newObject.get("metadata")).put("resourceVersion", resourceVersion);
+
+    return objectMapper.writeValueAsString(newObject);
   }
 }